Question2_DataPipeline/requirements.txt
# keep Airflow same as the image version
apache-airflow==2.7.3  

# compatible providers
apache-airflow-providers-openlineage>=1.8.0
apache-airflow-providers-apache-kafka

# kafka libs + data faker
kafka-python
confluent-kafka
faker
